requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
ffmpeg==1.4
ffprobe==0.5
ffprobe-python==1.0.3
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydub==0.25.1
requests==2.30.0
SpeechRecognition==3.10.0
urllib3==1.26.6
Werkzeug==2.3.4
zipp==3.15.0
easyocr==1.7.0

numpy~=1.24.3
opencv-python~=4.7.0.72
Pillow~=9.5.0